Refactoring
- moved all int64 to int (read previous comment)

+++ dh/DataHandling.go
@@ -13,11 +13,11 @@ import (
 
 // DiscrestiseVector takes a one-dimensional slice and discretises it.
 // Min/max and number of bins must be provided.
-func DiscrestiseVector(d []float64, bins int64, min, max float64) []int64 {
-	r := make([]int64, len(d), len(d))
+func DiscrestiseVector(d []float64, bins int, min, max float64) []int {
+	r := make([]int, len(d), len(d))
 	domain := max - min
 	for i := 0; i < len(d); i++ {
-		r[i] = int64(math.Min(((d[i]-min)/domain)*float64(bins), float64(bins-1)))
+		r[i] = int(math.Min(((d[i]-min)/domain)*float64(bins), float64(bins-1)))
 	}
 	return r
 }
@@ -26,15 +26,15 @@ func DiscrestiseVector(d []float64, bins int64, min, max float64) []int64 {
 // Min/max and number of bins must be provided for each column. The first
 // index of the data are the rows and the second index the columns, i.e.,
 // d[r][c] is the data point in the r-th row and c-th column
-func Discrestise(d [][]float64, bins []int64, min, max []float64) [][]int64 {
+func Discrestise(d [][]float64, bins []int, min, max []float64) [][]int {
 	rows := len(d)
-	ret := make([][]int64, rows, rows)
+	ret := make([][]int, rows, rows)
 	for r := 0; r < rows; r++ {
 		cols := len(d[r])
-		ret[r] = make([]int64, cols, cols)
+		ret[r] = make([]int, cols, cols)
 		for c := 0; c < cols; c++ {
 			domain := max[c] - min[c]
-			ret[r][c] = int64(math.Min(((d[r][c]-min[c])/domain)*float64(bins[c]), float64(bins[c]-1)))
+			ret[r][c] = int(math.Min(((d[r][c]-min[c])/domain)*float64(bins[c]), float64(bins[c]-1)))
 		}
 	}
 	return ret
@@ -42,13 +42,13 @@ func Discrestise(d [][]float64, bins []int64, min, max []float64) [][]int64 {
 
 // MakeUnivariate takes a two-dimensional discretised slice and returns
 // a one-dimensional representation of it.
-func MakeUnivariate(d [][]int64, bins []int64) []int64 {
+func MakeUnivariate(d [][]int, bins []int) []int {
 	rows := len(d)
 	cols := len(d[0])
-	ret := make([]int64, rows, rows)
+	ret := make([]int, rows, rows)
 
-	f := int64(1)
-	v := int64(0)
+	f := 1
+	v := 0
 	for r := 0; r < rows; r++ {
 		v = 0
 		f = 1
@@ -70,11 +70,11 @@ func MakeUnivariate(d [][]int64, bins []int64) []int64 {
 // a one-dimensional representation of it, in which only
 // consecutive values, i.e.,
 // [1,7,10,15] will be converted into [0,1,2,3]
-func Relabel(d []int64) []int64 {
+func Relabel(d []int) []int {
 	rows := len(d)
-	ret := make([]int64, rows, rows)
-	encountered := map[int64]bool{}
-	unique := []int64{}
+	ret := make([]int, rows, rows)
+	encountered := map[int]bool{}
+	unique := []int{}
 
 	for i, _ := range d {
 		if encountered[d[i]] == false {
@@ -87,7 +87,7 @@ func Relabel(d []int64) []int64 {
 	for r := 0; r < rows; r++ {
 		for i := 0; i < len(unique); i++ {
 			if unique[i] == d[r] {
-				ret[r] = int64(i)
+				ret[r] = i
 				break
 			}
 		}
@@ -137,7 +137,7 @@ func ReadData(filename string) (r [][]float64) {
 }
 
 // ExtractColumns [...]
-func ExtractColumns(data [][]float64, indices []int64) [][]float64 {
+func ExtractColumns(data [][]float64, indices []int) [][]float64 {
 
 	r := make([][]float64, len(data), len(data))
