+++ discrete/MorphologicalComputation.go
@@ -80,7 +80,7 @@ func MorphologicalComputationMI(pw2w1 [][]float64, pa1s1 [][]float64) float64 {
 // MorphologicalComputationSY quantifies morphological computation as the synergistic information that
 // W and A contain about W'. For more details, please read
 // TODO Paper reference
-func MorphologicalComputationSY(pw2w1a1 [][][]float64, iterations int64, eta bool) float64 {
+func MorphologicalComputationSY(pw2w1a1 [][][]float64, iterations int, eta bool) float64 {
 	split := IterativeScaling{}
 
 	split.NrOfVariables = 3
@@ -105,10 +105,10 @@ func MorphologicalComputationSY(pw2w1a1 [][][]float64, iterations int64, eta boo
 	var bar *pb.ProgressBar
 
 	if eta == true {
-		bar = pb.StartNew(int(iterations))
+		bar = pb.StartNew(iterations)
 	}
 
-	for i := 0; i < int(iterations); i++ {
+	for i := 0; i < iterations; i++ {
 		split.Iterate()
 		if eta == true {
 			bar.Increment()
@@ -126,7 +126,7 @@ func MorphologicalComputationSY(pw2w1a1 [][][]float64, iterations int64, eta boo
 // information that W and A contain about W', excluding the input distribution
 // (W,A). For more details, please read
 // TODO Paper reference
-func MorphologicalComputationSyNid(pw2w1a1 [][][]float64, iterations int64) float64 {
+func MorphologicalComputationSyNid(pw2w1a1 [][][]float64, iterations int) float64 {
 	split := IterativeScaling{}
 
 	split.NrOfVariables = 3
@@ -147,7 +147,7 @@ func MorphologicalComputationSyNid(pw2w1a1 [][][]float64, iterations int64) floa
 	split.Features["A,W'"] = []int{1, 2}
 
 	split.Init()
-	for i := 0; i < int(iterations); i++ {
+	for i := 0; i < iterations; i++ {
 		split.Iterate()
 	}
 
@@ -155,6 +155,6 @@ func MorphologicalComputationSyNid(pw2w1a1 [][][]float64, iterations int64) floa
 }
 
 // MorphologicalComputationP [...]
-func MorphologicalComputationP(pw2w1a1 [][][]float64, iterations int64, eta bool) float64 {
+func MorphologicalComputationP(pw2w1a1 [][][]float64, iterations int, eta bool) float64 {
 	return MorphologicalComputationW(pw2w1a1) - MorphologicalComputationSY(pw2w1a1, iterations, eta)
 }
