+++ discrete/state/ConditionalMutualInformation_test.go
@@ -11,10 +11,10 @@ import (
 
 func TestConditionalMutualInformation(t *testing.T) {
 	t.Log("Testing Conditional Mutual Information")
-	data := make([][]int64, 100)
+	data := make([][]int, 100)
 	for j := 0; j < 100; j++ {
 		for i := 0; i < 100; i++ {
-			data[i] = make([]int64, 3, 3)
+			data[i] = make([]int, 3, 3)
 			data[i][0] = rand.Int63n(100)
 			data[i][1] = rand.Int63n(100)
 			data[i][2] = rand.Int63n(100)
